# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Asesora-T API Doc",
        "description": "Documentación de la API de Asesora-T",
        "contact": {
            "email": "caamalmarco99@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://asesorat-backend.test:8080",
            "description": "SwaggerHUB API Mocking"
        }
    ],
    "paths": {
...
        "/api/alumnos/{idAlumno}/asesorias-confirmadas": {
            "get": {
                "tags": [
                    "Asesorias"
                ],
                "summary": "Obtener Asesorias Confirmadas",
                "description": "Devuelve las asesorias confirmadas del alumno",
                "operationId": "GetAllConfirmadas",
                "parameters": [
                    {
                        "name": "idAlumno",
                        "in": "path",
                        "description": "ID del alumno",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numero de página que por defecto es la 1",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginadorDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/asesorias/horarios-disponibles": {
            "get": {
                "tags": [
                    "Asesorias"
                ],
                "summary": "Obtener Asesorias Disponibles para el alumno",
                "description": "Devuelve las asesorias disponibles para el alumno",
                "operationId": "GetAllDisponibles",
                "parameters": [
                    {
                        "name": "horaInicio",
                        "in": "query",
                        "description": "Filtrar por Hora de Inicio",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "horaFin",
                        "in": "query",
                        "description": "Filtrar por Hora de Fin",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Filtrar por nombre del profesor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "turno",
                        "in": "query",
                        "description": "Filtrar por turno (0 = Matutino, 1 = Vespertino)",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "$ref": "#/components/schemas/TurnoEnum"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de la página actual que por defecto es 1",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginadorDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/asesorias/horarios-disponibles/cantidad-por-turnos": {
            "get": {
                "tags": [
                    "Asesorias"
                ],
                "summary": "Obtener Cantidad de Asesorias Disponibles por Turnos.",
                "description": "Devuelve el número de asesorias disponibles para el alumno por turnos.",
                "operationId": "GetCantidadAsesoriasDisponiblesPorTurnos",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "matutinos": {
                                            "type": "integer",
                                            "format": "int64"
                                        },
                                        "vespertinos": {
                                            "type": "integer",
                                            "format": "int64"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profesores/{profesorId}/asesorias/finalizadas": {
            "get": {
                "tags": [
                    "Asesorias"
                ],
                "summary": "Obtener lista paginada de asesorias finalizadas del profesor.",
                "description": "Devuelve una lista paginada de las asesorias finalizadas del profesor.",
                "operationId": "GetAllFinalizadasProfesor",
                "parameters": [
                    {
                        "name": "profesorId",
                        "in": "path",
                        "description": "ID del profesor",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de la página actual que por defecto es 1",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginadorDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profesores/{profesorId}/asesorias-confirmadas": {
            "get": {
                "tags": [
                    "Asesorias"
                ],
                "summary": "Obtener lista paginada de asesorias confirmadas de un profesor.",
                "description": "Devuelve una lista paginada de las asesorias confirmadas de un profesor.",
                "operationId": "GetAllConfirmadasProfesor",
                "parameters": [
                    {
                        "name": "profesorId",
                        "in": "path",
                        "description": "ID del profesor",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de la página actual que por defecto es 1",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginadorDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TurnoEnum": {
                "type": "integer",
                "format": "enum",
                "enum": [
                    0,
                    1
                ]
            },
...
            "PaginadorDTO": {
                "title": "Paginador",
                "properties": {
                    "current_page": {
                        "title": "Pagina Actual",
                        "description": "Página actual",
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "title": "Datos",
                        "description": "Datos que se obtenieron",
                        "type": "array",
                        "items": {
                            "oneOf": [
                                {
                                    "$ref": "#/components/schemas/ProfesorDTO"
                                }
                            ]
                        }
                    },
                    "first_page_url": {
                        "title": "URL ",
                        "description": "Url de la primera página",
                        "type": "string"
                    },
                    "from": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "last_page_url": {
                        "type": "string"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LinkPaginador"
                        }
                    },
                    "next_page_url": {
                        "type": "string"
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "prev_page_url": {
                        "type": "string"
                    },
                    "to": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}